Move mapper into utils.
